339.2. More on Route Params(Section 24: REST-Based React Apps)
And now to make sure that I see the correct component on the screen I'm going to click on streaming

on the top left side and then click on the edit button.

One more time.

And now I again see stream at it.

OK.

So we now know the exact ID of the stream that we're trying to edit.

Anytime that we click on that edit button.

So now the next thing I want to try to do is load up some information about that particular stream and

show it on the on the screen.

Remember eventually we want to allow a user to edit this information right we want to show the two text

inputs inside there so will probably make use of redux farm or something like that.

But whenever we go to the edit stream page we want to show the existing title and the existing description

of the stream as it currently is.

So we need to make sure that we've already loaded up the stream and that we can show some information

from it.

So like I said right now we'll just set a very early goal and say let's try to just print up the title

and description and not worry about setting up Reebok's form or anything like that.

So let's take a quick pause right here and we'll start to tackle that in the next video.