342.2 Fetching a Stream for Edit Stream(Section 24: REST-Based React Apps)
Essentially we just need to put in a little statement and say if we do not have our stream then return

like a div that says loading or something like that like so.

So if we have not yet loaded up our stream no problem where Id turn the div with voting.

Otherwise if we do have our stream attempt to print out the title so now we can save this again and

we'll see the name of the stream that I'm trying to edit which is I created distrain.

All right.

So that's pretty much going to take care of our data loading stuff.

Now everything you see inside a stream at it right now is going to essentially be repeated almost line

for line over to our other two components of stream delete and stream show because remember both stream

delete and stream show need to load up the appropriate stream in order to show some details about it

on the screen.

So for example when we try to show a window or a message to the user that says Are you sure you want

to delete that stream.

Will probably end up wanting to put like the title of this stream on there as well.

So we'll need to fetch the stream ahead of time and then for streams show.

Same thing very obvious in that case whenever a user is looking at a stream.

Yeah we need to load it up so we need to make sure that we wire up the streams show component to fetch

the data that it needs as well.

So again what you see inside of here is going to be just about duplicated into those other two components

as well in the future when we start to work with them.

All right.

So let's take a pause here now that we've got our data.

When we come back the next section we're going to start to put together the actual edit stream form.

And there's going to be a little fun twist around this component that we haven't quite spoken about

just yet so let's take a pause and we'll get to it in the next video.